+++ lib/options/word_list.dart
@@ -3,27 +3,36 @@ import 'package:path/path.dart' as path;
 class WordList {
   final String displayName;
   String fileName;
+  final int identifier;
 
   WordList({
     this.displayName,
     this.fileName,
+    this.identifier,
   }) {
-    fileName = path.join('assets', 'scowl_word_lists', this.fileName);
+    fileName = path.join(
+      'assets',
+      'scowl_word_lists',
+      this.fileName,
+    );
   }
 
   static final WordList _small = WordList(
     displayName: 'Simple',
     fileName: '35.txt',
+    identifier: 0,
   );
 
   static final WordList _default = WordList(
     displayName: 'Normal',
     fileName: '60.txt',
+    identifier: 1,
   );
 
   static final WordList _insane = WordList(
     displayName: 'Insane',
     fileName: '95.txt',
+    identifier: 2,
   );
 
   static List<WordList> _allWordLists = <WordList>[
@@ -38,5 +47,10 @@ class WordList {
 
   static WordList getDefault() {
     return _default;
+  }
+
+  static WordList getByIdentifier(int id) {
+    return getAllWordLists()
+        .firstWhere((WordList wordList) => wordList.identifier == id);
   }
 }